Refactor MM retriever
