+++ llama_index/core/__init__.py
@@ -1,4 +1,11 @@
+from llama_index.core.base_multi_modal_retriever import MultiModalRetriever
 from llama_index.core.base_query_engine import BaseQueryEngine
 from llama_index.core.base_retriever import BaseRetriever
+from llama_index.core.image_retriever import BaseImageRetriever
 
-__all__ = ["BaseRetriever", "BaseQueryEngine"]
+__all__ = [
+    "BaseRetriever",
+    "BaseQueryEngine",
+    "MultiModalRetriever",
+    "BaseImageRetriever",
+]
