+++ llama_index/core/base_multi_modal_retriever.py
@@ -2,9 +2,9 @@
 from abc import abstractmethod
 from typing import List
 
+from llama_index.core.image_retriever import BaseImageRetriever
 from llama_index.indices.base_retriever import BaseRetriever
 from llama_index.indices.query.schema import QueryType
-from llama_index.retrievers.image_retriever import BaseImageRetriever
 from llama_index.schema import NodeWithScore
