+++ llama_index/core/image_retriever.py
